+++ packages/admin/resources/views/livewire/components/products/index.blade.php
@@ -130,17 +130,16 @@
           <x-hub::table.cell colspan="24">
             @unless($selectAll)
               <span class="text-sm text-indigo-800">{{ __('adminhub::components.products.index.selected_products', [
-                'count' => 123,
-              ]); }} 
+                'count' => count($selected),
+              ]); }}
                 <strong>{{ $products->total() }}</strong>?</span>
               <button wire:click="selectAll" class="ml-1 text-blue-700 hover:underline">{{ __('adminhub::components.products.index.select_all_btn') }}</button>
             @else
-              <span class="text-sm text-indigo-800">{{ __('adminhub::components.products.index.you_have_selected_all', , [
-                'count' => 123,
-              ]); }} 
+              <span class="text-sm text-indigo-800">{{ __('adminhub::components.products.index.you_have_selected_all', [
+                'count' => $products->total(),
+              ]) }}
             </span>
             @endif
-
           </x-hub::table.cell>
         </x-hub::table.row>
       @endif