+++ packages/core/database/migrations/2022_05_10_100000_fix_last_four_on_transactions_table.php
@@ -1,0 +1,22 @@
+<?php
+
+use GetCandy\Base\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class FixLastFourOnTransactionsTable extends Migration
+{
+    public function up()
+    {
+        Schema::table($this->prefix.'transactions', function (Blueprint $table) {
+            $table->string('last_four', 4)->change();
+        });
+    }
+
+    public function down()
+    {
+        Schema::table($this->prefix.'transactions', function ($table) {
+            $table->smallInteger('last_four')->unsigned()->change();
+        });
+    }
+}