+++ packages/core/src/Base/Migration.php
@@ -20,4 +20,18 @@ abstract class Migration extends BaseMigration
     {
         $this->prefix = config('getcandy.database.table_prefix');
     }
+
+    /**
+     * Use the connection specified in config.
+     *
+     * @return void
+     */
+    public function getConnection()
+    {
+        if ($connection = config('getcandy.database.connection', false)) {
+            return $connection;
+        }
+
+        return parent::getConnection();
+    }
 }