+++ packages/core/CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+### Fixed
+
+- When adding a purchasable to the cart, a potential difference in key order for meta is taken into account.
+
+### Changed
+
 ## 2.0-beta12 - 2022-04-08
 
 ### Changed