+++ packages/core/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
@@ -6,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+
+- Added support to allow the scout driver to be defined per model.
+- `savedCart` relationship has been added to the `Cart` model.
 
 ### Fixed
 
@@ -184,12 +188,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## 2.0-beta2 - 2021-12-23
 
 ### Fixed
+
 - Default currency has `enabled` set to true.
 
 ### Changed
+
 - Install command no longer publishes hub assets
 
 ### Added
+
 - Added a default `CollectionGroup`.
 
 [View Changes](https://github.com/getcandy/core/compare/2.0-beta...2.0-beta2)