+++ packages/core/CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Incorrect price when currency has zero decimal places.
+
 ## 2.0-beta12 - 2022-04-08
 
 ### Changed